LPS-92618 regen more

+++ modules/apps/headless/headless-admin-user/headless-admin-user-client/src/main/java/com/liferay/headless/admin/user/client/serdes/v1_0/OrganizationSerDes.java
@@ -302,7 +302,7 @@ public class OrganizationSerDes {
 
 				if (jsonParserFieldValue != null) {
 					organization.setNumberOfOrganizations(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(
