+++ modules/apps/headless/headless-form/headless-form-client/src/main/java/com/liferay/headless/form/client/serdes/v1_0/FormDocumentSerDes.java
@@ -179,7 +179,7 @@ public class FormDocumentSerDes {
 			else if (Objects.equals(jsonParserFieldName, "sizeInBytes")) {
 				if (jsonParserFieldValue != null) {
 					formDocument.setSizeInBytes(
-						Long.valueOf((String)jsonParserFieldValue));
+						Integer.valueOf((String)jsonParserFieldValue));
 				}
 			}
 			else if (Objects.equals(jsonParserFieldName, "title")) {
